添加评论功能

+++ controllers/topic.go
@@ -15,7 +15,7 @@ func (c *TopicController) Get()  {
 	c.TplName = "topic.html"
 	c.Data["IsLogin"] = checkLogin(c.Ctx)
 
-	topics, err := models.GetAllTopics(false)
+	topics, err := models.GetAllTopics("", false)
 	if err != nil {
 		beego.Error(err)
 	}
@@ -26,28 +26,40 @@ func (c *TopicController) Post()  {
 	title := c.Input().Get("title")
 	content := c.Input().Get("content")
 	category := c.Input().Get("category")
-	titleId := c.Input().Get("titleId")
-	fmt.Println("titleId", titleId)
+	topicId := c.Input().Get("topicId")
+	fmt.Println("topicId", topicId)
 
 	var err error
-	if len(titleId)!=0 {  //如是参数里含有文章id，说明操作是修改文章
-		err = models.ModifyTopic(titleId, title, content)
+	if len(topicId)!=0 {  //如是参数里含有文章id，说明操作是修改文章
+		err = models.ModifyTopic(topicId, title, content, category)
+		c.Redirect("/topic", 302)
 	}else {
 		err = models.AddTopic(title, content, category)
+		c.Redirect("/topic", 302)
+
 	}
 
 	if err != nil {
 		beego.Error(err.Error())
 	}
-	c.Redirect("/", 302)
 }
 
 
 func(c *TopicController) Add()  {
+	if !checkLogin(c.Ctx) {       //检查是否需要重新登录
+		c.Redirect("/login", 302)
+		return
+	}
+
 	c.TplName = "topic_add.html"
 }
 
 func (c *TopicController) Delete()  {
+	if !checkLogin(c.Ctx) {       //检查是否需要重新登录
+		c.Redirect("/login", 302)
+		return
+	}
+
 	id := c.Ctx.Input.Param("0")  //通过自动路由获取值,如/topic/delete/2,相当于/topic?op=delete&&value=2
 	err := models.DeleteTopic(id)
 	if err != nil {
@@ -58,10 +70,15 @@ func (c *TopicController) Delete()  {
 }
 
 func (c *TopicController) Modify()  {
-	titleId := c.Ctx.Input.Param("0")
-	c.Data["TitleId"] = titleId
+	if !checkLogin(c.Ctx) {       //检查是否需要重新登录
+		c.Redirect("/login", 302)
+		return
+	}
 
-	topic, err := models.GetTopic(titleId)
+	topicId := c.Ctx.Input.Param("0")
+	c.Data["TopicId"] = topicId
+
+	topic, err := models.GetTopic(topicId)
 	if err!=nil {
 		beego.Error(err)
 	}
@@ -74,15 +91,26 @@ func (c *TopicController) Modify()  {
 func (c *TopicController) View()  {
 	c.Data["IsTopic"] = true
 	c.TplName = "topic_view.html"
+	c.Data["IsLogin"] = checkLogin(c.Ctx)
 
-	titleId := c.Ctx.Input.Param("0")
-	fmt.Println("View",titleId)
-	topic, err := models.GetTopic(titleId)
+	topicId := c.Ctx.Input.Param("0")
+	fmt.Println("View", topicId)
+	topic, err := models.GetTopic(topicId)
+
 	if err != nil {
 		beego.Error(err)
 		c.Redirect("/", 302)
 		return
 	}
 	c.Data["Topic"] = topic
+
+	replies, err := models.GetAllReplies(topicId)
+	if err != nil {
+		beego.Error(err)
+		c.Redirect("/", 302)
+		return
+	}
+
+	c.Data["Replies"] = replies
 
 }
